Move general diagnostics data types into their own module
